scripts/requirements.txt
macapptree
Pillow